* Added MS Fabric as output plugin * Refactored ADX and EH plugin to reuse common code. * Added corresponding test cases
