+++ plugins/outputs/all/microsoft_fabric.go
@@ -1,0 +1,3 @@
+package all
+
+import _ "github.com/influxdata/telegraf/plugins/outputs/microsoft_fabric"
